+++ changelog.md
@@ -1,4 +1,7 @@
 # Changelog
+## Unreleased
+* Breaking: Add an optional `cancellationToken` parameter to all async `IQueryable` extensions. (polyadic/funcky#433)
+* Breaking: Move extensions to `Funcky.Extensions` namespace. (polyadic/funcky#460)
 
 ## 0.1.0
 Initial release